+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/FlutterLocalNotificationsPlugin.java
@@ -19,6 +19,8 @@ import android.media.RingtoneManager;
 import android.net.Uri;
 import android.os.Build.VERSION;
 import android.os.Build.VERSION_CODES;
+import android.os.Handler;
+import android.os.Looper;
 import android.service.notification.StatusBarNotification;
 import android.text.Html;
 import android.text.Spanned;
@@ -70,6 +72,10 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 import io.flutter.FlutterInjector;
 import io.flutter.embedding.engine.loader.FlutterLoader;
@@ -141,10 +147,18 @@ public class FlutterLocalNotificationsPlugin
   private static final String CANCEL_TAG = "tag";
   static String NOTIFICATION_DETAILS = "notificationDetails";
   static Gson gson;
+  static ExecutorService executor;
+  static Handler handler;
   private MethodChannel channel;
   private Context applicationContext;
   private Activity mainActivity;
   private Intent launchIntent;
+
+  public FlutterLocalNotificationsPlugin() {
+    executor =
+        new ThreadPoolExecutor(0, 1, 30L, TimeUnit.SECONDS, new LinkedBlockingQueue<Runnable>());
+    handler = new Handler(Looper.getMainLooper());
+  }
 
   @SuppressWarnings("deprecation")
   public static void registerWith(io.flutter.plugin.common.PluginRegistry.Registrar registrar) {
@@ -154,18 +168,18 @@ public class FlutterLocalNotificationsPlugin
     plugin.onAttachedToEngine(registrar.context(), registrar.messenger());
   }
 
-  static void rescheduleNotifications(Context context) {
+  static void rescheduleNotifications(Context context, Result result) {
     initAndroidThreeTen(context);
     ArrayList<NotificationDetails> scheduledNotifications = loadScheduledNotifications(context);
     for (NotificationDetails scheduledNotification : scheduledNotifications) {
       if (scheduledNotification.repeatInterval == null) {
         if (scheduledNotification.timeZoneName == null) {
-          scheduleNotification(context, scheduledNotification, false);
+          scheduleNotification(context, scheduledNotification, false, result);
         } else {
-          zonedScheduleNotification(context, scheduledNotification, false);
+          zonedScheduleNotification(context, scheduledNotification, false, result);
         }
       } else {
-        repeatNotification(context, scheduledNotification, false);
+        repeatNotification(context, scheduledNotification, false, result);
       }
     }
   }
@@ -334,35 +348,34 @@ public class FlutterLocalNotificationsPlugin
   }
 
   private static void saveScheduledNotifications(
-      Context context, ArrayList<NotificationDetails> scheduledNotifications) {
+      Context context, ArrayList<NotificationDetails> scheduledNotifications, Result result) {
     Gson gson = buildGson();
     String json = gson.toJson(scheduledNotifications);
     SharedPreferences sharedPreferences =
         context.getSharedPreferences(SCHEDULED_NOTIFICATIONS, Context.MODE_PRIVATE);
     SharedPreferences.Editor editor = sharedPreferences.edit();
     editor.putString(SCHEDULED_NOTIFICATIONS, json);
-    tryCommittingInBackground(editor, 3);
+    commitAsync(editor, result);
   }
 
-  private static void tryCommittingInBackground(
-      final SharedPreferences.Editor editor, final int tries) {
-    if (tries == 0) {
-      return;
-    }
-    new Thread(
-            new Runnable() {
-              @Override
-              public void run() {
-                final boolean isCommitted = editor.commit();
-                if (!isCommitted) {
-                  tryCommittingInBackground(editor, tries - 1);
-                }
-              }
-            })
-        .start();
+  private static void commitAsync(final SharedPreferences.Editor editor, final Result result) {
+    executor.execute(
+        new Runnable() {
+          @Override
+          public void run() {
+            final boolean committed = editor.commit();
+            handler.post(
+                new Runnable() {
+                  @Override
+                  public void run() {
+                    result.success(committed);
+                  }
+                });
+          }
+        });
   }
 
-  static void removeNotificationFromCache(Context context, Integer notificationId) {
+  static void removeNotificationFromCache(Context context, Integer notificationId, Result result) {
     ArrayList<NotificationDetails> scheduledNotifications = loadScheduledNotifications(context);
     for (Iterator<NotificationDetails> it = scheduledNotifications.iterator(); it.hasNext(); ) {
       NotificationDetails notificationDetails = it.next();
@@ -371,7 +384,7 @@ public class FlutterLocalNotificationsPlugin
         break;
       }
     }
-    saveScheduledNotifications(context, scheduledNotifications);
+    saveScheduledNotifications(context, scheduledNotifications, result);
   }
 
   @SuppressWarnings("deprecation")
@@ -389,7 +402,8 @@ public class FlutterLocalNotificationsPlugin
   private static void scheduleNotification(
       Context context,
       final NotificationDetails notificationDetails,
-      Boolean updateScheduledNotificationsCache) {
+      Boolean updateScheduledNotificationsCache,
+      Result result) {
     Gson gson = buildGson();
     String notificationDetailsJson = gson.toJson(notificationDetails);
     Intent notificationIntent = new Intent(context, ScheduledNotificationReceiver.class);
@@ -413,14 +427,15 @@ public class FlutterLocalNotificationsPlugin
     }
 
     if (updateScheduledNotificationsCache) {
-      saveScheduledNotification(context, notificationDetails);
+      saveScheduledNotification(context, notificationDetails, result);
     }
   }
 
   private static void zonedScheduleNotification(
       Context context,
       final NotificationDetails notificationDetails,
-      Boolean updateScheduledNotificationsCache) {
+      Boolean updateScheduledNotificationsCache,
+      Result result) {
     Gson gson = buildGson();
     String notificationDetailsJson = gson.toJson(notificationDetails);
     Intent notificationIntent = new Intent(context, ScheduledNotificationReceiver.class);
@@ -448,12 +463,12 @@ public class FlutterLocalNotificationsPlugin
     }
 
     if (updateScheduledNotificationsCache) {
-      saveScheduledNotification(context, notificationDetails);
+      saveScheduledNotification(context, notificationDetails, result);
     }
   }
 
   static void scheduleNextRepeatingNotification(
-      Context context, NotificationDetails notificationDetails) {
+      Context context, NotificationDetails notificationDetails, Result result) {
     long repeatInterval = calculateRepeatIntervalMilliseconds(notificationDetails);
     long notificationTriggerTime =
         calculateNextNotificationTrigger(notificationDetails.calledAt, repeatInterval);
@@ -466,15 +481,7 @@ public class FlutterLocalNotificationsPlugin
     AlarmManager alarmManager = getAlarmManager(context);
     AlarmManagerCompat.setExactAndAllowWhileIdle(
         alarmManager, AlarmManager.RTC_WAKEUP, notificationTriggerTime, pendingIntent);
-    saveScheduledNotification(context, notificationDetails);
-  }
-
-  private static PendingIntent getActivityPendingIntent(Context context, int id, Intent intent) {
-    int flags = PendingIntent.FLAG_UPDATE_CURRENT;
-    if (VERSION.SDK_INT >= VERSION_CODES.M) {
-      flags |= PendingIntent.FLAG_IMMUTABLE;
-    }
-    return PendingIntent.getActivity(context, id, intent, flags);
+    saveScheduledNotification(context, notificationDetails, result);
   }
 
   private static PendingIntent getBroadcastPendingIntent(Context context, int id, Intent intent) {
@@ -488,7 +495,8 @@ public class FlutterLocalNotificationsPlugin
   private static void repeatNotification(
       Context context,
       NotificationDetails notificationDetails,
-      Boolean updateScheduledNotificationsCache) {
+      Boolean updateScheduledNotificationsCache,
+      Result result) {
     long repeatInterval = calculateRepeatIntervalMilliseconds(notificationDetails);
 
     long notificationTriggerTime = notificationDetails.calledAt;
@@ -524,7 +532,7 @@ public class FlutterLocalNotificationsPlugin
           AlarmManager.RTC_WAKEUP, notificationTriggerTime, repeatInterval, pendingIntent);
     }
     if (updateScheduledNotificationsCache) {
-      saveScheduledNotification(context, notificationDetails);
+      saveScheduledNotification(context, notificationDetails, result);
     }
   }
 
@@ -560,7 +568,7 @@ public class FlutterLocalNotificationsPlugin
   }
 
   private static void saveScheduledNotification(
-      Context context, NotificationDetails notificationDetails) {
+      Context context, NotificationDetails notificationDetails, Result result) {
     ArrayList<NotificationDetails> scheduledNotifications = loadScheduledNotifications(context);
     ArrayList<NotificationDetails> scheduledNotificationsToSave = new ArrayList<>();
     for (NotificationDetails scheduledNotification : scheduledNotifications) {
@@ -570,7 +578,7 @@ public class FlutterLocalNotificationsPlugin
       scheduledNotificationsToSave.add(scheduledNotification);
     }
     scheduledNotificationsToSave.add(notificationDetails);
-    saveScheduledNotifications(context, scheduledNotificationsToSave);
+    saveScheduledNotifications(context, scheduledNotificationsToSave, result);
   }
 
   private static int getDrawableResourceId(Context context, String name) {
@@ -1032,25 +1040,25 @@ public class FlutterLocalNotificationsPlugin
   }
 
   static void zonedScheduleNextNotification(
-      Context context, NotificationDetails notificationDetails) {
+      Context context, NotificationDetails notificationDetails, Result result) {
     initAndroidThreeTen(context);
     String nextFireDate = getNextFireDate(notificationDetails);
     if (nextFireDate == null) {
       return;
     }
     notificationDetails.scheduledDateTime = nextFireDate;
-    zonedScheduleNotification(context, notificationDetails, true);
+    zonedScheduleNotification(context, notificationDetails, true, result);
   }
 
   static void zonedScheduleNextNotificationMatchingDateComponents(
-      Context context, NotificationDetails notificationDetails) {
+      Context context, NotificationDetails notificationDetails, Result result) {
     initAndroidThreeTen(context);
     String nextFireDate = getNextFireDateMatchingDateTimeComponents(notificationDetails);
     if (nextFireDate == null) {
       return;
     }
     notificationDetails.scheduledDateTime = nextFireDate;
-    zonedScheduleNotification(context, notificationDetails, true);
+    zonedScheduleNotification(context, notificationDetails, true, result);
   }
 
   static String getNextFireDate(NotificationDetails notificationDetails) {
@@ -1197,7 +1205,10 @@ public class FlutterLocalNotificationsPlugin
   }
 
   @Override
-  public void onDetachedFromEngine(FlutterPluginBinding binding) {}
+  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+    handler.removeCallbacksAndMessages(null);
+    executor.shutdown();
+  }
 
   @Override
   public void onAttachedToActivity(ActivityPluginBinding binding) {
@@ -1316,16 +1327,14 @@ public class FlutterLocalNotificationsPlugin
     Map<String, Object> arguments = call.arguments();
     Integer id = (Integer) arguments.get(CANCEL_ID);
     String tag = (String) arguments.get(CANCEL_TAG);
-    cancelNotification(id, tag);
-    result.success(null);
+    cancelNotification(id, tag, result);
   }
 
   private void repeat(MethodCall call, Result result) {
     Map<String, Object> arguments = call.arguments();
     NotificationDetails notificationDetails = extractNotificationDetails(result, arguments);
     if (notificationDetails != null) {
-      repeatNotification(applicationContext, notificationDetails, true);
-      result.success(null);
+      repeatNotification(applicationContext, notificationDetails, true, result);
     }
   }
 
@@ -1333,8 +1342,7 @@ public class FlutterLocalNotificationsPlugin
     Map<String, Object> arguments = call.arguments();
     NotificationDetails notificationDetails = extractNotificationDetails(result, arguments);
     if (notificationDetails != null) {
-      scheduleNotification(applicationContext, notificationDetails, true);
-      result.success(null);
+      scheduleNotification(applicationContext, notificationDetails, true, result);
     }
   }
 
@@ -1346,8 +1354,7 @@ public class FlutterLocalNotificationsPlugin
         notificationDetails.scheduledDateTime =
             getNextFireDateMatchingDateTimeComponents(notificationDetails);
       }
-      zonedScheduleNotification(applicationContext, notificationDetails, true);
-      result.success(null);
+      zonedScheduleNotification(applicationContext, notificationDetails, true, result);
     }
   }
 
@@ -1389,8 +1396,7 @@ public class FlutterLocalNotificationsPlugin
         applicationContext.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE);
     SharedPreferences.Editor editor = sharedPreferences.edit();
     editor.putString(DEFAULT_ICON, defaultIcon);
-    tryCommittingInBackground(editor, 3);
-    result.success(true);
+    commitAsync(editor, result);
   }
 
   /// Extracts the details of the notifications passed from the Flutter side and also validates that
@@ -1486,7 +1492,7 @@ public class FlutterLocalNotificationsPlugin
         && !isValidDrawableResource(applicationContext, icon, result, INVALID_ICON_ERROR_CODE);
   }
 
-  private void cancelNotification(Integer id, String tag) {
+  private void cancelNotification(Integer id, String tag, Result result) {
     Intent intent = new Intent(applicationContext, ScheduledNotificationReceiver.class);
     PendingIntent pendingIntent = getBroadcastPendingIntent(applicationContext, id, intent);
     AlarmManager alarmManager = getAlarmManager(applicationContext);
@@ -1497,7 +1503,7 @@ public class FlutterLocalNotificationsPlugin
     } else {
       notificationManager.cancel(tag, id);
     }
-    removeNotificationFromCache(applicationContext, id);
+    removeNotificationFromCache(applicationContext, id, result);
   }
 
   private void cancelAllNotifications(Result result) {
@@ -1518,8 +1524,7 @@ public class FlutterLocalNotificationsPlugin
       alarmManager.cancel(pendingIntent);
     }
 
-    saveScheduledNotifications(applicationContext, new ArrayList<NotificationDetails>());
-    result.success(null);
+    saveScheduledNotifications(applicationContext, new ArrayList<NotificationDetails>(), result);
   }
 
   @Override